+++ libs/components/src/lib/navigate-back-button/navigate-back-button.component.ts
@@ -8,7 +8,7 @@ import {
   templateUrl:     './navigate-back-button.component.html',
   styleUrls:       [ './navigate-back-button.component.scss' ],
   changeDetection: ChangeDetectionStrategy.OnPush,
-  host:            { class: 'srm-navigate-back-button' }
+  host: { class: 'rxap-navigate-back-button' }
 })
 export class NavigateBackButtonComponent {