+++ libs/xml-parser/CHANGELOG.md
@@ -3,6 +3,39 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [12.1.0-next.1](https://gitlab.com/rxap/packages/compare/@rxap/xml-parser@12.1.0-next.0...@rxap/xml-parser@12.1.0-next.1) (2021-07-20)
+
+
+### Features
+
+* **element-record:** add propertyKey attribute ([56e101e](https://gitlab.com/rxap/packages/commit/56e101e02d29ceb879bb4a8ee3372728eca0a317))
+
+
+
+
+
+# [12.1.0-next.0](https://gitlab.com/rxap/packages/compare/@rxap/xml-parser@12.0.10-next.0...@rxap/xml-parser@12.1.0-next.0) (2021-07-20)
+
+
+### Features
+
+* **element-record:** add raw option ([fb3ca9e](https://gitlab.com/rxap/packages/commit/fb3ca9e2bff7f0361f5f4f67d21225f9c3c75a68))
+
+
+
+
+
+## [12.0.10-next.0](https://gitlab.com/rxap/packages/compare/@rxap/xml-parser@12.0.9...@rxap/xml-parser@12.0.10-next.0) (2021-07-20)
+
+
+### Bug Fixes
+
+* **element-child-text-content:** only fail required if property is not defined ([def9653](https://gitlab.com/rxap/packages/commit/def9653b8c526f83e523672b75aed93950d6f3b5))
+
+
+
+
+
 ## [12.0.9](https://gitlab.com/rxap/packages/compare/@rxap/xml-parser@12.0.9-next.0...@rxap/xml-parser@12.0.9) (2021-07-09)
 
 **Note:** Version bump only for package @rxap/xml-parser